requirements.txt
h5py==3.15.1
numpy==2.4.0
pandas==2.3.3
scikit-learn==1.6.0
tqdm==4.67.1
joblib==1.5.3
httplib2==0.31.0
google_api_python_client==2.185.0
google_auth==2.41.1
zstandard==0.25.0